moved redis create client to util and fixed its imports. Renamed init to initRoom
