+++ backend/src/util.ts
@@ -1,0 +1,11 @@
+import { createClient } from "redis";
+
+export async function CreateRedisClient() {
+
+    let client = await createClient({
+        url: "redis://redis"
+    }).connect();
+
+    return client;
+
+}
